refactor : seperate CategorySetup & CategoryList - reason : for more readability
